Handle move and copy constructors
We make the object non-copyable, because to copy one of these objects
would have to result in a double delete. But we do allow moving. In the
wrapped pointer case we transfer ownership. In the wrapped unique_ptr
case we transfer the borrowed pointer and leave the moved-from object
holding a null pointer.