Use getCh timeout for timeout in main loop
The result is that updates from keypresses will be handled with no
delay, but otherwise the refresh rate is determined by the 50 ms
blocking time set for getCh. This answers the question in our
previous comment, but handling the timeout and interrupts is done
by the libraries for us under the hood.